+++ backend/web/src/frontend/src/main.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import VueGtag from './analytics/index.js';
+import vueGtag from './plugins/vue-gtag.js';
 import App from './App.vue';
 import router from './router/index.js';
 import store from './store/index.js';
@@ -8,7 +8,7 @@ import vuetify from './plugins/vuetify.js';
 Vue.config.productionTip = false;
 
 new Vue({
-  VueGtag,
+  vueGtag,
   router,
   store,
   vuetify,
